darc_backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
mysqlclient==2.2.8
PyMySQL==1.1.2
sqlparse==0.5.5
djangorestframework==3.14.0
django-filter==24.1
django-cors-headers==4.3.1
python-decouple==3.8
